Move ClockSync to util submodule
