+++ src/ezmsg/lsl/inlet.py
@@ -10,6 +10,8 @@ import numpy as np
 import numpy.typing as npt
 import pylsl
 
+from .util import ClockSync
+
 
 fmt2npdtype = {
     pylsl.cf_double64: float,  # Prefer native type for float64
@@ -74,43 +76,6 @@ class LSLInletState(ez.State):
     resolver: typing.Optional[pylsl.ContinuousResolver] = None
     inlet: typing.Optional[pylsl.StreamInlet] = None
     clock_offset: float = 0.0
-
-
-class ClockSync:
-    def __init__(self, alpha: float = 0.1, min_interval: float = 0.5):
-        self.alpha = alpha
-        self.min_interval = min_interval
-
-        self.offset = 0.0
-        self.last_update = 0.0
-        self.count = 0
-
-    async def update(self, force: bool = False, burst: int = 4) -> None:
-        dur_since_last = time.time() - self.last_update
-        dur_until_next = self.min_interval - dur_since_last
-        if force or dur_until_next <= 0:
-            offsets = []
-            for ix, _ in enumerate(range(burst)):
-                if (self.count + ix) % 2:
-                    y, x = time.time(), pylsl.local_clock()
-                else:
-                    x, y = pylsl.local_clock(), time.time()
-                # TODO: Use adaptive linear fit instead of simple subtraction.
-                offsets.append(y - x)
-                self.last_update = y
-                await asyncio.sleep(0.001)
-            offset = np.mean(offsets)
-
-            if self.count > 0:
-                # Exponential decay smoothing
-                offset = (1 - self.alpha) * self.offset + self.alpha * offset
-            self.offset = offset
-            self.count += burst
-        else:
-            await asyncio.sleep(dur_until_next)
-
-    def convert_timestamp(self, lsl_timestamp: float) -> float:
-        return lsl_timestamp + self.offset
 
 
 class LSLInletUnit(ez.Unit):
